finally, finished booking layout
