+++ app/_layout.tsx
@@ -3,8 +3,6 @@ import AuthProvider from "@/context/AuthProvider";
 import { Stack } from "expo-router";
 import { setStatusBarStyle } from "expo-status-bar";
 import { TamaguiProvider } from "tamagui";
-import { PortalProvider } from "tamagui";
-import { useFonts } from "expo-font";
 
 import config from "../tamagui.config";
 
@@ -24,14 +22,12 @@ export default function RootLayout() {
 
   return (
     <TamaguiProvider config={config}>
-      <PortalProvider shouldAddRootHost>
-        <AuthProvider>
-          <Stack>
-            <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-            <Stack.Screen name="(auth)" options={{ headerShown: false }} />
-          </Stack>
-        </AuthProvider>
-      </PortalProvider>
+      <AuthProvider>
+        <Stack>
+          <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+          <Stack.Screen name="(auth)" options={{ headerShown: false }} />
+        </Stack>
+      </AuthProvider>
     </TamaguiProvider>
   );
 }
